Move function.js to utils dir
